exception handler delete

+++ src/main/java/com/joaogabriel/vendas/service/UserService.java
@@ -4,43 +4,54 @@ import java.util.List;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Service;
 
 import com.joaogabriel.vendas.entities.User;
 import com.joaogabriel.vendas.repositories.UserRepository;
+import com.joaogabriel.vendas.service.exceptions.DatabaseException;
 import com.joaogabriel.vendas.service.exceptions.ResourceNotFoundException;
 
 @Service
 public class UserService {
-	
+
 	@Autowired
 	private UserRepository userRepository;
-	
-	public List<User> findAll(){
+
+	public List<User> findAll() {
 		return userRepository.findAll();
 	}
-	
+
 	public User findById(Long id) {
 		Optional<User> user = userRepository.findById(id);
-		
+
 		return user.orElseThrow(() -> new ResourceNotFoundException(id));
 	}
-	
+
 	public User insert(User obj) {
 		return userRepository.save(obj);
 	}
-	
+
 	public void delete(Long id) {
-		userRepository.deleteById(id);
+		try {
+			
+			if(!userRepository.existsById(id)) throw new ResourceNotFoundException(id);
+			userRepository.deleteById(id);
+		}catch(ResourceNotFoundException e) {
+			throw new ResourceNotFoundException(id);
+		}catch(DataIntegrityViolationException e ) {
+			throw new DatabaseException(e.getMessage());
+		}
 	}
-	
+
 	public User update(Long id, User obj) {
-		
+
 		User user = userRepository.getReferenceById(id);
 		updateData(user, obj);
 		return userRepository.save(user);
 	}
-	
+
 	public void updateData(User user, User obj) {
 		user.setName(obj.getName());
 		user.setEmail(obj.getEmail());
